requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
psycopg2-binary==2.9.8
python-dotenv==1.0.0
flask-cors==5.0.0
marshmallow-SQLAlchemy==1.1.0
Flask Authlib